Rename namesapce and enable replay mode
