+++ ouster-ros/include/ouster_ros/os_ros.h
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2018-2022, Ouster, Inc.
+ * Copyright (c) 2018-2023, Ouster, Inc.
  * All rights reserved.
  *
  * @file os_ros.h
