+++ ouster-ros/launch/sensor.composite.launch.py
@@ -17,7 +17,7 @@ def generate_launch_description():
             composable_node_descriptions=[
                 ComposableNode(
                     package='ouster_ros',
-                    plugin='nodelets_os::OusterSensor',
+                    plugin='ouster_ros::OusterSensor',
                     name='os_sensor',
                     parameters=[
                         {'sensor_hostname' : 'os-122151001683.local'},
@@ -31,7 +31,7 @@ def generate_launch_description():
                 ),
                 ComposableNode(
                     package='ouster_ros',
-                    plugin='nodelets_os::OusterCloud',
+                    plugin='ouster_ros::OusterCloud',
                     name='os_cloud',
                     parameters=[
                         {'tf_prefix' : ' '},
@@ -39,7 +39,7 @@ def generate_launch_description():
                 ),
                 ComposableNode(
                     package='ouster_ros',
-                    plugin='nodelets_os::OusterImage',
+                    plugin='ouster_ros::OusterImage',
                     name='os_image'
                 )
             ],
