+++ ouster-ros/src/os_replay_node.cpp
@@ -1,38 +1,38 @@
 /**
- * Copyright (c) 2018-2022, Ouster, Inc.
+ * Copyright (c) 2018-2023, Ouster, Inc.
  * All rights reserved.
  *
- * @file os_replay_nodelet.cpp
- * @brief This nodelet mainly handles publishing saved metadata
+ * @file os_replay_node.cpp
+ * @brief This node mainly handles publishing saved metadata
  *
  */
 
-#include "ouster_ros/os_client_base_nodelet.h"
+#include "ouster_ros/visibility_control.h"
+#include "ouster_ros/os_client_base_node.h"
+
 
 namespace sensor = ouster::sensor;
 
-namespace nodelets_os {
+namespace ouster_ros {
 
 class OusterReplay : public OusterClientBase {
    public:
-    COMPOSITION_PUBLIC
-    
-    explicit OusterSensor(const rclcpp::NodeOptions & options)
-    : Node("os_replay", options)
-    {
+    OUSTER_ROS_PUBLIC
+    explicit OusterReplay(const rclcpp::NodeOptions& options)
+    : OusterClientBase("os_replay", options) {
         onInit();
     }
 
    private:
     virtual void onInit() override {
-        auto& pnh = getPrivateNodeHandle();
-        auto meta_file = pnh.param("metadata", std::string{});
+        declare_parameter("metadata", rclcpp::PARAMETER_STRING);
+        auto meta_file = get_parameter("metadata").as_string();
         if (!is_arg_set(meta_file)) {
-            NODELET_ERROR("Must specify metadata file in replay mode");
+            RCLCPP_ERROR(get_logger(), "Must specify metadata file in replay mode");
             throw std::runtime_error("metadata no specificed");
         }
 
-        NODELET_INFO("Running in replay mode");
+        RCLCPP_INFO(get_logger(), "Running in replay mode");
 
         // populate info for config service
         try {
@@ -40,15 +40,16 @@ class OusterReplay : public OusterClientBase {
             cached_metadata = to_string(info);
             display_lidar_info(info);
         } catch (const std::runtime_error& e) {
-            NODELET_ERROR("Error when running in replay mode: %s", e.what());
+            RCLCPP_ERROR(get_logger(),
+            "Error when running in replay mode: %s", e.what());
         }
 
         OusterClientBase::onInit();
     }
 };
 
-}  // namespace nodelets_os
+}  // namespace ouster_ros
 
 #include <rclcpp_components/register_node_macro.hpp>
 
-RCLCPP_COMPONENTS_REGISTER_NODE(nodelets_os::OusterReplay)
+RCLCPP_COMPONENTS_REGISTER_NODE(ouster_ros::OusterReplay)
